requirements.txt
yfinance==0.2.54
pandas==2.2.3
numpy>=1.26.0,<2.1
matplotlib>=3.7.0,<3.10
requests==2.32.3
python-dotenv==1.1.0
jquants-api-client
jupyter==1.1.1
streamlit>=1.35.0
plotly>=5.22.0
